requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
protobuf>=3.20,<5.0
numpy>=1.24.0
line-bot-sdk>=2.0.0
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
flask>=2.0.0
